requirements.txt
playwright==1.49.0
pynput==1.7.7
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
opencv-python>=4.8.0
scikit-image>=0.21.0



